requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
blobfile==3.0.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comtypes==1.4.9
distro==1.9.0
filelock==3.17.0
fire==0.7.0
frozenlist==1.5.0
fsspec==2024.12.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.28.1
idna==3.10
Jinja2==3.1.5
jiter==0.8.2
llama_models==0.1.0
llama_stack==0.1.0
llama_stack_client==0.1.0
llamaapi==0.1.36
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.2
openai==1.60.2
packaging==24.2
pandas==2.2.3
pefile==2023.2.7
pillow==11.1.0
prompt_toolkit==3.0.50
propcache==0.2.1
protobuf==5.29.3
pyaml==25.1.0
PyAudio==0.2.14
pycryptodomex==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2025.1
pypiwin32==223
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyttsx3==2.98
pytz==2025.1
pywin32==308
pywin32-ctypes==0.2.3
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
safetensors==0.5.2
sentencepiece==0.2.0
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
SpeechRecognition==3.14.1
sympy==1.13.1
termcolor==2.5.0
tiktoken==0.8.0
tokenizers==0.21.0
torch==2.6.0
tqdm==4.67.1
transformers==4.48.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wcwidth==0.2.13
yarl==1.18.3